# perf: optimized JSON batch handling