--- a/route/batched_event.go
+++ b/route/batched_event.go
@@ -3,12 +3,16 @@ package route
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"time"
 
 	"github.com/honeycombio/refinery/config"
 	"github.com/honeycombio/refinery/types"
 	"github.com/tinylib/msgp/msgp"
+	"github.com/valyala/fastjson"
 )
+
+var fastJsonParserPool fastjson.ParserPool
 
 type batchedEvent struct {
 	Timestamp           string        `json:"time"`
@@ -35,6 +39,8 @@ func (b *batchedEvent) getSampleRate() uint {
 }
 
 func (b *batchedEvent) UnmarshalJSON(data []byte) error {
+	// This method is now primarily used as a fallback for individual event unmarshaling
+	// Most of the time, events are unmarshaled via batchedEvents.UnmarshalJSON
 	type tempEvent struct {
 		Timestamp  string          `json:"time"`
 		SampleRate int64           `json:"samplerate"`
@@ -42,25 +48,23 @@ func (b *batchedEvent) UnmarshalJSON(data []byte) error {
 	}
 
 	var temp tempEvent
-	err := json.Unmarshal(data, &temp)
-	if err != nil {
+	if err := json.Unmarshal(data, &temp); err != nil {
 		return err
 	}
 
-	// Copy the simple fields
 	b.Timestamp = temp.Timestamp
 	b.SampleRate = temp.SampleRate
-
-	// Initialize Data with config
 	b.Data = types.NewPayload(b.cfg, nil)
 
-	// Convert JSON Data to MessagePack and use optimized unmarshaling
+	// Convert data field to MessagePack and use optimized unmarshaling
 	buf := httpBodyBufferPool.Get().(*bytes.Buffer)
 	defer recycleHTTPBodyBuffer(buf)
 
 	msgpackData, err := types.JSONToMessagePack(buf.Bytes()[:0], temp.Data)
+	if err != nil {
+		return err
+	}
 
-	// Use the same optimized unmarshaling logic as UnmarshalMsg
 	_, err = b.coreFieldsExtractor.UnmarshalPayload(msgpackData, &b.Data)
 	return err
 }
@@ -167,18 +171,30 @@ func (b *batchedEvents) MarshalJSON() ([]byte, error) {
 }
 
 func (b *batchedEvents) UnmarshalJSON(data []byte) error {
-	var rawEvents []json.RawMessage
-	err := json.Unmarshal(data, &rawEvents)
+	parser := fastJsonParserPool.Get()
+	defer fastJsonParserPool.Put(parser)
+
+	v, err := parser.ParseBytes(data)
 	if err != nil {
 		return err
 	}
 
-	b.events = make([]batchedEvent, len(rawEvents))
-	for i := range b.events {
+	if v.Type() != fastjson.TypeArray {
+		return fmt.Errorf("expected JSON array")
+	}
+
+	arr, err := v.Array()
+	if err != nil {
+		return err
+	}
+
+	b.events = make([]batchedEvent, len(arr))
+	for i, eventValue := range arr {
 		b.events[i].cfg = b.cfg
 		b.events[i].coreFieldsExtractor = b.coreFieldsExtractor
 
-		err = json.Unmarshal(rawEvents[i], &b.events[i])
+		// Parse each event directly using fastjson
+		err = b.unmarshalBatchedEventFromFastJSON(&b.events[i], eventValue)
 		if err != nil {
 			return err
 		}
@@ -186,3 +202,66 @@ func (b *batchedEvents) UnmarshalJSON(data []byte) error {
 
 	return nil
 }
+
+// unmarshalBatchedEventFromFastJSON populates a batchedEvent from a fastjson.Value
+func (b *batchedEvents) unmarshalBatchedEventFromFastJSON(event *batchedEvent, v *fastjson.Value) error {
+	if v.Type() != fastjson.TypeObject {
+		return fmt.Errorf("expected JSON object for event")
+	}
+
+	obj, err := v.Object()
+	if err != nil {
+		return err
+	}
+
+	// Initialize Data with config
+	event.Data = types.NewPayload(event.cfg, nil)
+
+	var dataValue *fastjson.Value
+
+	// Visit each field in the event object
+	obj.Visit(func(key []byte, v *fastjson.Value) {
+		if err != nil {
+			return
+		}
+
+		switch string(key) {
+		case "time":
+			if v.Type() == fastjson.TypeString {
+				s, _ := v.StringBytes()
+				event.Timestamp = string(s)
+			}
+		case "samplerate":
+			if v.Type() == fastjson.TypeNumber {
+				event.SampleRate = v.GetInt64()
+			}
+		case "data":
+			if v.Type() == fastjson.TypeObject {
+				dataValue = v
+			}
+		}
+	})
+
+	if err != nil {
+		return err
+	}
+
+	// Convert data field to MessagePack and use optimized unmarshaling
+	if dataValue != nil {
+		buf := httpBodyBufferPool.Get().(*bytes.Buffer)
+		defer recycleHTTPBodyBuffer(buf)
+
+		// Use AppendJSONValue directly to avoid expensive MarshalTo() call
+		msgpackData, err := types.AppendJSONValue(buf.Bytes()[:0], dataValue)
+		if err != nil {
+			return err
+		}
+
+		// Use the same optimized unmarshaling logic as UnmarshalMsg
+		_, err = event.coreFieldsExtractor.UnmarshalPayload(msgpackData, &event.Data)
+		return err
+	}
+
+	return nil
+}
+