## Which problem is this PR solving?
JSON is still a thing, so we need to handle it efficiently.

## Short description of the changes
Short-circuits the batch JSON parsing pathway to directly pick out the fields we want using optimized code, and then turn the data blob directly into messagepack. That makes this slower than messagepack, but faster than OTLP, which isn't bad.

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkRouterBatch/batch_json-12     346258        26287         -92.41%

benchmark                              old allocs     new allocs     delta
BenchmarkRouterBatch/batch_json-12     1935           30             -98.45%

benchmark                              old bytes     new bytes     delta
BenchmarkRouterBatch/batch_json-12     113436        17539         -84.54%
```